+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 Added:
 
+- added JWT token auth
+- added login functionality
 - added pull and push error messages
 - added push and pull cmd; 17 feb 2026
 - added makefile based compilation